Replace abstract PlayerSpawnMenu network message with an  interface since Editor throws  warning at non-sealed GameNetworkMessage, and abstract class can't be sealed........
